Autologout fix

+++ frontend/src/util/AutoLogout.js
@@ -1,13 +1,20 @@
 import axiosConfig from "./axiosConfig";
+import Cookies from "js-cookie";
+import swal from "sweetalert";
 
 export const autoLogout = async () => {
-
-        try {
-          await axiosConfig.post("/user/logout")
-          localStorage.clear()
-          console.log("automatically logged out");
-        } catch (error) {
-          console.log(error);
-    }
-}
- 
+  const isLoggedToken = Cookies.get("isLogged");
+  const expiresInMs = isLoggedToken - new Date().getTime();
+  const localItems = localStorage.getItem("userId");
+  if ((!isLoggedToken && localItems) || (expiresInMs <= 0 && localItems)) {
+    await axiosConfig.post("/user/logout");
+    localStorage.clear();
+    swal({
+      title: "Du wurdest automatisch abgemeldet.",
+      icon: "warning",
+      dangerMode: true,
+    }).then(() => {
+      location.reload()
+    });
+  }
+};
